+++ lib/ui/components/components.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:wave/config.dart';
 import 'package:wave/wave.dart';
 
-import '../domain/models/game_item.dart';
+import '../../domain/models/game_item.dart';
 
 class Headers extends StatelessWidget {
   final String? title;
@@ -281,45 +281,6 @@ class CenterLandscape extends StatelessWidget {
           ),
         if (isLandscape) const Spacer(flex: 1),
       ],
-    );
-  }
-}
-
-class ThemeSwitch extends StatelessWidget {
-  final Widget? child;
-  final bool isDark;
-  final VoidCallback? onToggle;
-
-  const ThemeSwitch({
-    Key? key,
-    this.isDark = false,
-    this.child,
-    this.onToggle,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final child = this.child;
-    return Material(
-      child: Stack(
-        children: [
-          if (child != null) child,
-          Align(
-            alignment: Alignment.topRight,
-            child: SafeArea(
-              child: Padding(
-                padding: const EdgeInsets.all(12.0),
-                child: IconButton(
-                  icon: Icon(
-                    isDark ? Icons.wb_sunny_outlined : Icons.nightlight_round,
-                  ),
-                  onPressed: () => onToggle?.call(),
-                ),
-              ),
-            ),
-          )
-        ],
-      ),
     );
   }
 }
